config/requirements.txt
marshmallow-sqlalchemy==0.28.1
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0